feat: add ETag support for caching in report responses
